[refactor] #17 MemberController 및 JwtAuthenticationFilter 리팰터링
